+++ LCLandmineOutside/Patches/RoundManagerPatch.cs
@@ -3,27 +3,45 @@ using System.Collections.Generic;
 using UnityEngine;
 using System.Linq;
 using Unity.Netcode;
+using LCHazardsOutside.Abstract;
+using Dissonance.Integrations.Unity_NFGO;
+using System.Collections;
+using System;
 
 namespace LCHazardsOutside.Patches {
 
     [HarmonyPatch(typeof(RoundManager))]
     internal class RoundManagerPatch {
 
-        private const string LAYER_NAME = "MapHazards";
+        private const string HAZARD_LAYER_NAME = "MapHazards";
 
         [HarmonyPatch(nameof(RoundManager.SpawnOutsideHazards))]
         [HarmonyPrefix]
         static void SpawnHazardsOutsidePatch(RoundManager __instance) {
-            if (__instance.currentLevel.sceneName == "CompanyBuilding" || !__instance.IsServer || !__instance.IsHost) {
+            if (!__instance.currentLevel.spawnEnemiesAndScrap || !__instance.IsServer || !__instance.IsHost) {
                 return;
             }
 
             // Tie hazard spawn to seed.
-            LCHazardsOutside.GetLogger().LogDebug("randomMapSeed: " + StartOfRound.Instance.randomMapSeed);
+            Plugin.GetLogger().LogDebug("randomMapSeed: " + StartOfRound.Instance.randomMapSeed);
             System.Random random = new(StartOfRound.Instance.randomMapSeed + 587);
-            List<GameObject> spawnDenialPoints = [.. GameObject.FindGameObjectsWithTag("SpawnDenialPoint")];
 
-            LCHazardsOutside.GetLogger().LogDebug("Getting spawnable objects..");
+            int noHazardSpawnChance = Plugin.instance.noHazardSpawnChance.Value;
+            if (noHazardSpawnChance > 0)
+            {
+                double chance = noHazardSpawnChance / 100.0;
+
+                if (random.NextDouble() < chance)
+                {
+                    Plugin.GetLogger().LogDebug("No hazards spawned outside due to global chance.");
+                    return;
+                }
+            }
+
+            string sanitizedPlanetName = LCUtils.GetNumberlessPlanetName(__instance.currentLevel);
+            Plugin.GetLogger().LogDebug($"Planetname: {sanitizedPlanetName}");
+
+            List<GameObject> spawnDenialPoints = [.. GameObject.FindGameObjectsWithTag("SpawnDenialPoint")];
             List<SpawnableMapObject> hazardObjects = [.. __instance.currentLevel.spawnableMapObjects];
 
             if (hazardObjects.Any()) {
@@ -32,66 +50,69 @@ namespace LCHazardsOutside.Patches {
                     SpawnableMapObject hazardObject = hazardObjects[i];
                     bool isIncreasedMapHazardSpawn = __instance.increasedMapHazardSpawnRateIndex == i;
 
-                    LCHazardsOutside.GetLogger().LogDebug("Current spawnable object: " + hazardObject.prefabToSpawn.ToString());
+                    Plugin.GetLogger().LogDebug("Current spawnable object: " + hazardObject.prefabToSpawn);
 
                     bool isTurret = hazardObject.prefabToSpawn.GetComponentInChildren<Turret>() != null;
                     bool isLandmine = hazardObject.prefabToSpawn.GetComponentInChildren<Landmine>() != null;
 
-                    LCHazardsOutside.GetLogger().LogDebug("Is Turret: " + isTurret);
-                    LCHazardsOutside.GetLogger().LogDebug("Is Landmine: " + isLandmine);
-
-                    if (isLandmine && LCHazardsOutside.instance.configEnableLandmine.Value)
+                    if (isLandmine && Plugin.instance.enableLandmine.Value)
                     {
-                        LCHazardsOutside.GetLogger().LogInfo("Spawning landmines outside...");
-                        int configMinSpawnRate = LCHazardsOutside.instance.configLandmineMinSpawnRate.Value;
-                        int configMaxSpawnRate = LCHazardsOutside.instance.configLandmineMaxSpawnRate.Value;
+                        Plugin.GetLogger().LogInfo("Spawning landmines outside...");
+                        int configMinSpawnRate = Plugin.instance.globalLandmineMinSpawnRate.Value;
+                        int configMaxSpawnRate = Plugin.instance.globalLandmineMaxSpawnRate.Value;
+                        Plugin.instance.landmineMoonMap.TryGetValue(sanitizedPlanetName, out MoonMinMax moonMinMax);
 
-                        determineMinMaxSpawnRates(isIncreasedMapHazardSpawn, configMinSpawnRate, configMaxSpawnRate, out int minSpawnRate, out int maxSpawnRate);
+                        DetermineMinMaxSpawnRates(isIncreasedMapHazardSpawn, configMinSpawnRate, configMaxSpawnRate, moonMinMax, out int minSpawnRate, out int maxSpawnRate);
                         HazardCalculationContainer landmineContainer = new(random, spawnDenialPoints, hazardObject, minSpawnRate, maxSpawnRate);
 
-                        CalculateHazardSpawn(__instance, landmineContainer);
+                        CalculateHazardSpawn(__instance, landmineContainer, new DefaultSpawnStrategy());
                     }  
 
-                    if (isTurret && LCHazardsOutside.instance.configEnableTurret.Value)
+                    if (isTurret && Plugin.instance.enableTurret.Value)
                     {
-                        LCHazardsOutside.GetLogger().LogInfo("Spawning turrets outside...");
-                        int configMinSpawnRate = LCHazardsOutside.instance.configTurretMinSpawnRate.Value;
-                        int configMaxSpawnRate = LCHazardsOutside.instance.configTurretMaxSpawnRate.Value;
+                        Plugin.GetLogger().LogInfo("Spawning turrets outside...");
+                        int configMinSpawnRate = Plugin.instance.globalTurretMinSpawnRate.Value;
+                        int configMaxSpawnRate = Plugin.instance.globalTurretMaxSpawnRate.Value;
+                        Plugin.instance.turretMoonMap.TryGetValue(sanitizedPlanetName, out MoonMinMax moonMinMax);
 
-                        determineMinMaxSpawnRates(false, configMinSpawnRate, configMaxSpawnRate, out int minSpawnRate, out int maxSpawnRate);
+                        DetermineMinMaxSpawnRates(false, configMinSpawnRate, configMaxSpawnRate, moonMinMax, out int minSpawnRate, out int maxSpawnRate);
                         HazardCalculationContainer turretContainer = new(random, spawnDenialPoints, hazardObject, minSpawnRate, maxSpawnRate)
                         {
                             NeedsSafetyZone = true,
                             SpawnRatioMultiplier = 2f
                         };
 
-                        CalculateHazardSpawn(__instance, turretContainer);   
+                        CalculateHazardSpawn(__instance, turretContainer, new DefaultSpawnStrategy());   
                     }
                           
                     // Unknown Hazard
-                    if (!isLandmine && !isTurret && LCHazardsOutside.instance.configEnableCustomHazard.Value)
+                    if (!isLandmine && !isTurret && Plugin.instance.enableCustomHazard.Value)
                     {
-                        LCHazardsOutside.GetLogger().LogInfo("Spawning custom hazards outside...");
-                        int configMinSpawnRate = LCHazardsOutside.instance.configCustomHazardMinSpawnRate.Value;
-                        int configMaxSpawnRate = LCHazardsOutside.instance.configCustomHazardMaxSpawnRate.Value;
+                        Plugin.GetLogger().LogInfo("Spawning custom hazards outside...");
+                        int configMinSpawnRate = Plugin.instance.globalCustomHazardMinSpawnRate.Value;
+                        int configMaxSpawnRate = Plugin.instance.globalCustomHazardMaxSpawnRate.Value;
+                        Plugin.instance.customHazardMoonMap.TryGetValue(sanitizedPlanetName, out MoonMinMax moonMinMax);
 
-                        determineMinMaxSpawnRates(isIncreasedMapHazardSpawn, configMinSpawnRate, configMaxSpawnRate, out int minSpawnRate, out int maxSpawnRate);
+                        DetermineMinMaxSpawnRates(isIncreasedMapHazardSpawn, configMinSpawnRate, configMaxSpawnRate, moonMinMax, out int minSpawnRate, out int maxSpawnRate);
                         HazardCalculationContainer customContainer = new(random, spawnDenialPoints, hazardObject, 1, 1);
 
-                        CalculateHazardSpawn(__instance, customContainer);
+                        CalculateHazardSpawn(__instance, customContainer, new DefaultSpawnStrategy());
                     }
                 }
             }
 
-            LCHazardsOutside.GetLogger().LogInfo("Outside hazard spawning done.");
+            Plugin.GetLogger().LogInfo("Outside hazard spawning done.");
         }
 
-        private static void determineMinMaxSpawnRates(bool increasedMapHazardSpawnRate, int configMinSpawnRate, int configMaxSpawnRate, out int minSpawnRate, out int maxSpawnRate)
+        private static void DetermineMinMaxSpawnRates(bool increasedMapHazardSpawnRate, int configMinSpawnRate, int configMaxSpawnRate, MoonMinMax moonMinMax, out int minSpawnRate, out int maxSpawnRate)
         {
+            int effectiveUserMinValue = moonMinMax != null ? moonMinMax.min : configMinSpawnRate;
+            int effectiveUserMaxValue = moonMinMax != null ? moonMinMax.max : configMaxSpawnRate;
+
             // Range from 0 to max config.
-            minSpawnRate = Mathf.Min(Mathf.Max(configMinSpawnRate, 0), configMaxSpawnRate);
+            minSpawnRate = Mathf.Min(Mathf.Max(effectiveUserMinValue, 0), effectiveUserMaxValue);
             // Range from min to 50.
-            maxSpawnRate = Mathf.Max(Mathf.Min(configMaxSpawnRate, 50), minSpawnRate);
+            maxSpawnRate = Mathf.Max(Mathf.Min(effectiveUserMaxValue, 50), minSpawnRate);
 
             if (increasedMapHazardSpawnRate)
             {
@@ -100,7 +121,7 @@ namespace LCHazardsOutside.Patches {
             }
         }
 
-        private static void CalculateHazardSpawn(RoundManager __instance, HazardCalculationContainer hazardCalculationContainer)
+        private static void CalculateHazardSpawn(RoundManager __instance, HazardCalculationContainer hazardCalculationContainer, SpawnStrategy spawnStrategy)
         {
             Transform[] shipSpawnPathPoints = __instance.shipSpawnPathPoints;
 
@@ -110,25 +131,31 @@ namespace LCHazardsOutside.Patches {
             Transform shipLandingTransform = shipSpawnPathPoints.Last();
 
             int actualSpawnRate = hazardCalculationContainer.Random.Next(hazardCalculationContainer.MinSpawnRate, hazardCalculationContainer.MaxSpawnRate);
-            LCHazardsOutside.GetLogger().LogDebug("Actual spawn rate: " + actualSpawnRate);
+            Plugin.GetLogger().LogDebug("Actual spawn rate: " + actualSpawnRate);
 
             List<Vector3> safetyPositions = new(2);
 
             Vector3 shipLandPosition = shipLandingTransform.position;
             Vector3 mainEntrancePosition = RoundManager.FindMainEntrancePosition(false, true);
-            safetyPositions.Add(shipLandPosition);
-            LCHazardsOutside.GetLogger().LogDebug("Ship spawn point: " + shipLandPosition);
 
-            CalculateCenterPosition(shipLandPosition, mainEntrancePosition, hazardCalculationContainer.SpawnRatioMultiplier, out Vector3 middlePoint, out float spawnRadius);
+            safetyPositions.Add(shipLandPosition);
+            Plugin.GetLogger().LogDebug("Ship spawn point: " + shipLandPosition);
+
+            spawnStrategy.CalculateCenterPosition(shipLandPosition, mainEntrancePosition, [], hazardCalculationContainer.SpawnRatioMultiplier, out Vector3 middlePoint, out float spawnRadius);
+
+            List<GameObject> gameObjects = [];
+
+            int roomLayerMask = 1 << LayerMask.NameToLayer("Room");
+            int effectiveLayerMask = LCUtils.IsVanillaMoon(__instance.currentLevel) ? roomLayerMask : -1;
 
             for (int j = 0; j < actualSpawnRate; j++)
             {
                 System.Random random = hazardCalculationContainer.Random;
-                (Vector3 randomPosition, Quaternion quaternion) = SpawnHelper.GetRandomGroundPosition(middlePoint, spawnRadius, hazardCalculationContainer.Random, -1);
+                (Vector3 randomPosition, Quaternion quaternion) = spawnStrategy.GetRandomGroundPositionAndRotation(middlePoint, spawnRadius, hazardCalculationContainer.Random, effectiveLayerMask);
 
-                if (randomPosition == middlePoint)
+                if (randomPosition == Vector3.zero)
                 {
-                    LCHazardsOutside.GetLogger().LogDebug("No NavMesh hit!");
+                    Plugin.GetLogger().LogDebug("No NavMesh hit!");
                     continue;
                 }
 
@@ -137,35 +164,35 @@ namespace LCHazardsOutside.Patches {
                 // Do not spawn the hazard if it's too close to a spawn denial point.
                 if (invalidSpawnPointFound)
                 {
-                    LCHazardsOutside.GetLogger().LogDebug("Hazard was too close to denial or safety zone and was therefore deleted: " + randomPosition);
+                    Plugin.GetLogger().LogDebug("Hazard was too close to denial or safety zone and was therefore deleted: " + randomPosition);
                     continue;
                 }
 
-                SpawnHazard(__instance, random, hazardCalculationContainer.SpawnableMapObject, randomPosition, quaternion);
+                gameObjects.Add(InstantiateHazardObject(__instance, random, hazardCalculationContainer.SpawnableMapObject, randomPosition, quaternion));
 
                 hazardCounter++;
             }
 
-            LCHazardsOutside.GetLogger().LogDebug("Total hazard amount: " + hazardCounter);
+            __instance.StartCoroutine(SpawnHazardsInBulk(gameObjects));
+
+            Plugin.GetLogger().LogDebug("Total hazard amount: " + hazardCounter);
         }
 
-        private static void SpawnHazard(RoundManager __instance, System.Random random, SpawnableMapObject spawnableMapObject, Vector3 position, Quaternion quaternion)
+        private static GameObject InstantiateHazardObject(RoundManager __instance, System.Random random, SpawnableMapObject spawnableMapObject, Vector3 position, Quaternion quaternion)
         {
-            LCHazardsOutside.GetLogger().LogDebug("Spawn hazard outside at: " + position);
+            Plugin.GetLogger().LogDebug("Spawn hazard outside at: " + position);
             GameObject gameObject = UnityEngine.Object.Instantiate(spawnableMapObject.prefabToSpawn, position, quaternion, __instance.mapPropsContainer.transform);
 
             if (spawnableMapObject.spawnFacingAwayFromWall)
             {
                 gameObject.transform.eulerAngles = new Vector3(0f, __instance.YRotationThatFacesTheFarthestFromPosition(position + Vector3.up * 0.2f), 0f);
             }
-            else
-            {
-                gameObject.transform.eulerAngles = new Vector3(gameObject.transform.eulerAngles.x, random.Next(0, 360), gameObject.transform.eulerAngles.z);
-            }
 
             gameObject.SetActive(value: true);
-            gameObject.layer = LayerMask.NameToLayer(LAYER_NAME);
-            gameObject.GetComponent<NetworkObject>().Spawn(destroyWithScene: true);
+            gameObject.layer = LayerMask.NameToLayer(HAZARD_LAYER_NAME);
+            //gameObject.GetComponent<NetworkObject>().Spawn(destroyWithScene: true);
+
+            return gameObject;
         }
 
         private static bool IsInvalidSpawnPoint(List<GameObject> spawnDenialPoints, Transform[] shipPathPoints, Vector3 randomPosition)
@@ -212,11 +239,34 @@ namespace LCHazardsOutside.Patches {
             return false;
         }
 
-        private static void CalculateCenterPosition(Vector3 shipLandPosition, Vector3 mainEntrancePosition, float spawnRadiusMultiplier, out Vector3 centerPosition, out float spawnRadius)
+        public static IEnumerator SpawnHazardsInBulk(List<GameObject> gameObjects)
         {
-            centerPosition = (shipLandPosition + mainEntrancePosition) / 2;
-            spawnRadius = Vector3.Distance(mainEntrancePosition, centerPosition) * spawnRadiusMultiplier;
-            centerPosition.y = Mathf.Max(shipLandPosition.y, mainEntrancePosition.y);
+            yield return new WaitWhile(() => Plugin.instance.IsCoroutineRunning);
+
+            Plugin.instance.IsCoroutineRunning = true;
+            Plugin.GetLogger().LogDebug($"SpawnHazardsInBulk Coroutine running.");
+            const int bulkSize = 10;
+            for (int i = 0; i < gameObjects.Count; i += bulkSize)
+            {
+                int range = Mathf.Min(bulkSize, gameObjects.Count - i);
+                for (int j = 0; j < range; j++)
+                {
+                    GameObject objectToSpawn = gameObjects[i + j];
+                    try
+                    {
+                        objectToSpawn.GetComponent<NetworkObject>().Spawn(destroyWithScene: true);
+                    }
+                    catch (Exception e)
+                    {
+                        Plugin.GetLogger().LogError($"NetworkObject could not be spawned: {e}");
+                    }
+                }
+
+                yield return new WaitForSeconds(0.5f);
+            }
+
+            Plugin.instance.IsCoroutineRunning = false;
+            Plugin.GetLogger().LogDebug($"SpawnHazardsInBulk Coroutine done.");
         }
     }
 }