+++ LCLandmineOutside/VanillaMoon.cs
@@ -1,0 +1,14 @@
+﻿namespace LCHazardsOutside
+{
+    internal enum VanillaMoon
+    {
+        experimentation,
+        assurance,
+        offense,
+        vow,
+        march,
+        rend,
+        dine,
+        titan
+    }
+}